move 'obj_size' to be a stand-alone function
