+++ electrumsv/util/cache.py
@@ -1,14 +1,10 @@
-import bitcoinx
-from collections import deque
-from itertools import chain
 import sys
-from sys import getsizeof
 from threading import RLock
 from typing import Dict, List, Optional, Tuple
 
-from ..transaction import Transaction, XPublicKey, XTxOutput, XTxInput
-from ..constants import MAXIMUM_TXDATA_CACHE_SIZE_MB, MINIMUM_TXDATA_CACHE_SIZE_MB, \
-    ScriptType
+from .misc import obj_size
+from ..transaction import Transaction
+from ..constants import MAXIMUM_TXDATA_CACHE_SIZE_MB, MINIMUM_TXDATA_CACHE_SIZE_MB
 
 
 class Node:
@@ -79,11 +75,11 @@ class LRUCache:
                 assert value != old_value, "duplicate set not supported"
                 previous_node.next = next_node
                 next_node.previous = previous_node
-                self.current_size -= self.obj_size(old_value)
+                self.current_size -= obj_size(old_value)
                 del self._cache[key]
                 removals.append((key, old_value))
 
-            size = self.obj_size(value)
+            size = obj_size(value)
             if value is not None and size <= self._max_size:
                 added_node = self._add(key, value, size)
                 added = True
@@ -118,69 +114,7 @@ class LRUCache:
                 node.previous, node.next, node.key, node.value
             previous_node.next = next_node
             next_node.previous = previous_node
-            self.current_size -= self.obj_size(discard_value)
+            self.current_size -= obj_size(discard_value)
             del self._cache[discard_key]
             removals.append((discard_key, discard_value))
         return removals
-
-    def obj_size(self, o):
-        """This is a modified version of: https://code.activestate.com/recipes/577504/
-        to suit our bitcoin-specific needs
-
-        Returns the approximate memory footprint of an object and all of its contents.
-
-        Automatically finds the contents of the following builtin containers and
-        their subclasses:  tuple, list, deque, dict, set.
-
-        Additionally calculates the size of:
-        - electrumsv.transaction.Transaction
-        - electrumsv.transaction.XTxInput
-        - electrumsv.transaction.XTxOutput
-        - electrumsv.constants.ScriptType
-        - bitcoinx.Script
-        - bitcoinx.XPublicKey - approximation based on their serialized bytes footprint
-        """
-        dict_handler = lambda d: chain.from_iterable(d.items())
-
-        def attrs_object_iterator(obj):
-            """This is for iterating over attributes on classes produced via the 3rd
-            party library "attrs"""
-            return (getattr(obj, field.name) for field in obj.__attrs_attrs__)
-
-        all_handlers = {
-            tuple: iter,
-            list: iter,
-            deque: iter,
-            dict: dict_handler,
-            set: iter,
-            Transaction: attrs_object_iterator,
-            XTxInput: attrs_object_iterator,
-            XTxOutput: attrs_object_iterator}
-
-        seen = set()  # track which object id's have already been seen
-        default_size = getsizeof(0)  # estimate sizeof object without __sizeof__
-
-        def sizeof(o):
-
-            if id(o) in seen:  # do not double count the same object
-                return 0
-            seen.add(id(o))
-            s = getsizeof(o, default_size)
-
-            if isinstance(o, bitcoinx.Script):
-                s = len(o)
-
-            if isinstance(o, ScriptType):
-                s = 28
-
-            if isinstance(o, XPublicKey):
-                s = len(o.to_bytes())  # easiest approximation
-
-            for typ, handler in all_handlers.items():
-                if isinstance(o, typ):
-                    s += sum(map(sizeof, handler(o)))
-                    break
-
-            return s
-
-        return sizeof(o)
